:sparkles: rename staging_sync to chart_sync
