+++ apps/chart_sync/cli.py
@@ -15,7 +15,7 @@ from slack_sdk import WebClient
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 
-from apps.staging_sync.admin_api import AdminAPI
+from apps.chart_sync.admin_api import AdminAPI
 from apps.wizard.pages.chart_diff.chart_diff import ChartDiffModified
 from apps.wizard.utils.env import OWIDEnv, get_container_name
 from etl import config
@@ -136,7 +136,7 @@ def cli(
     """
     if _is_commit_sha(source):
         source = _get_git_branch_from_commit_sha(source)
-        log.info("staging_sync.use_branch", branch=source)
+        log.info("chart_sync.use_branch", branch=source)
 
     source_engine = OWIDEnv.from_staging_or_env_file(source).get_engine()
     target_engine = OWIDEnv.from_staging_or_env_file(target).get_engine()
@@ -154,7 +154,7 @@ def cli(
             else:
                 chart_ids = _modified_chart_ids_by_admin(source_session)
 
-            log.info("staging_sync.start", n=len(chart_ids), chart_ids=chart_ids)
+            log.info("chart_sync.start", n=len(chart_ids), chart_ids=chart_ids)
 
             for chart_id in chart_ids:
                 diff = ChartDiffModified.from_chart_id(chart_id, source_session, target_session)
@@ -168,7 +168,7 @@ def cli(
                 # Exclude charts with variables whose catalogPath matches the provided string
                 if not _matches_include_exclude(diff.source_chart, source_session, include, exclude):
                     log.info(
-                        "staging_sync.skip",
+                        "chart_sync.skip",
                         slug=chart_slug,
                         reason="filtered by --include/--exclude",
                         chart_id=chart_id,
@@ -183,7 +183,7 @@ def cli(
                     # Configs are equal, no need to update
                     if diff.configs_are_equal():
                         log.info(
-                            "staging_sync.skip",
+                            "chart_sync.skip",
                             slug=diff.target_chart.config["slug"],
                             reason="identical chart already exists",
                             chart_id=chart_id,
@@ -194,7 +194,7 @@ def cli(
                     if chartdiff:
                         # Change has been approved, update the chart
                         if diff.approved:
-                            log.info("staging_sync.chart_update", slug=chart_slug, chart_id=chart_id)
+                            log.info("chart_sync.chart_update", slug=chart_slug, chart_id=chart_id)
                             if not dry_run:
                                 target_api.update_chart(chart_id, diff.source_chart.config)
 
@@ -202,7 +202,7 @@ def cli(
                         # Rejected chart diff
                         # elif diff.rejected:
                         #     log.info(
-                        #         "staging_sync.rejected",
+                        #         "chart_sync.rejected",
                         #         slug=chart_slug,
                         #         chart_id=chart_id,
                         #     )
@@ -213,7 +213,7 @@ def cli(
                         # approving because of change in prod?
                         elif diff.unapproved:
                             log.warning(
-                                "staging_sync.unapproved_chart_update",
+                                "chart_sync.unapproved_chart_update",
                                 slug=chart_slug,
                                 chart_id=chart_id,
                                 source_updatedAt=str(diff.source_chart.updatedAt),
@@ -233,7 +233,7 @@ def cli(
                         # warn if chart has been updated in production after the staging server got created
                         if diff.target_chart.updatedAt > min(staging_created_at, diff.source_chart.updatedAt):
                             log.warning(
-                                "staging_sync.chart_modified_in_target",
+                                "chart_sync.chart_modified_in_target",
                                 slug=chart_slug,
                                 target_updatedAt=str(diff.target_chart.updatedAt),
                                 source_updatedAt=str(diff.source_chart.updatedAt),
@@ -253,7 +253,7 @@ def cli(
                         # chart hasn't been updated in production, update it directly
                         if approve_revisions and revs and staging_created_at > diff.target_chart.updatedAt:
                             log.info(
-                                "staging_sync.update_chart",
+                                "chart_sync.update_chart",
                                 slug=chart_slug,
                                 chart_id=chart_id,
                             )
@@ -294,14 +294,14 @@ def cli(
                                     # chart revision already exists
                                     target_session.rollback()
                                     log.info(
-                                        "staging_sync.skip",
+                                        "chart_sync.skip",
                                         reason="revision already exists",
                                         slug=chart_slug,
                                         chart_id=chart_id,
                                     )
                                     continue
 
-                            log.info("staging_sync.create_chart_revision", slug=chart_slug, chart_id=chart_id)
+                            log.info("chart_sync.create_chart_revision", slug=chart_slug, chart_id=chart_id)
 
                 # Chart is new, create it
                 else:
@@ -319,7 +319,7 @@ def cli(
                         # Rejected chart diff
                         # elif diff.rejected:
                         #     log.info(
-                        #         "staging_sync.rejected",
+                        #         "chart_sync.rejected",
                         #         slug=chart_slug,
                         #         chart_id=chart_id,
                         #     )
@@ -328,7 +328,7 @@ def cli(
                         # Not approved, create the chart, but notify us about it
                         elif diff.unapproved:
                             log.warning(
-                                "staging_sync.create_unapproved_chart",
+                                "chart_sync.create_unapproved_chart",
                                 slug=chart_slug,
                                 chart_id=chart_id,
                             )
@@ -351,7 +351,7 @@ def cli(
                             resp = {"chartId": None}
 
                     log.info(
-                        "staging_sync.create_chart",
+                        "chart_sync.create_chart",
                         published=diff.source_chart.config.get("isPublished"),
                         slug=chart_slug,
                         new_chart_id=resp["chartId"],
