+++ apps/chart_sync/instructions.md
