+++ apps/cli/__init__.py
@@ -180,7 +180,7 @@ GROUPS = (
         {
             "name": "Charts",
             "commands": {
-                "chart-sync": "apps.staging_sync.cli.cli",
+                "chart-sync": "apps.chart_sync.cli.cli",
                 "chart-gpt": "etl.chart_revision.v2.chartgpt.cli",
                 "chart-upgrade": "etl.chart_revision.cli.main_cli",
             },
