+++ apps/wizard/config/config.yml
@@ -102,8 +102,8 @@ sections:
         emoji: "🌟"
         image_url: "https://camo.githubusercontent.com/38a295d2c16cd880446f874a786e94fb168d6aadfdcbc4c8b7dd45c8337b6d1d/68747470733a2f2f6f7572776f726c64696e646174612e6f72672f677261706865722f6578706f7274732f6c6966652d657870656374616e63792e737667"
       - title: "Chart Sync"
-        alias: staging_sync
-        entrypoint: ../staging_sync/app.py
+        alias: chart-sync
+        entrypoint: ../chart_sync/app.py
         description: Sync charts between servers
         maintainer: "@mojmir"
         emoji: "🔄"
@@ -158,7 +158,7 @@ sections:
       disable:
         production: True
     - title: "Chart Diff"
-      alias: chart_diff
+      alias: chart-diff
       entrypoint: pages/chart_diff/app.py
       description: Chart differences
       maintainer: "@mojmir"
